+++ cmig_tools_utils/matlab/plotQQ.m
@@ -1,0 +1,118 @@
+function [fig, ll] = plotQQ(h_ax, pVals, strTitle, colOver)
+% Function that reads a GWAS summary file and creates a QQ plot
+%% Inputs:
+% fileName:     full path to a GWAS summary file
+%
+% outDir:       full path to where the figure should be saved (if empty, 
+%               figure handle is returned and image is not saved)
+%
+% strTitle:     character indicating the title of the figure (if left
+%               empty, no title will be put on top of the figure); this is
+%               also used for deciding output name; if empty, the output
+%               plot name is 'QQPlot.png'; else 'QQ_strTitle'
+%
+% fileFlag:     indicate which software was used to generate the GWAS 
+%               summary file (not case sensitive):
+%                   * 'FEMA'    (default)
+%                   * 'GCTA'    (assumes fastGWA output)
+%                   * 'Regenie'
+%                   * 'PLINK'   (assumes PLINK 2)
+%
+% h_ax:         handle to a figure axis that should be used for plotting
+%               (optional input; use this for creating plots that should 
+%               be added to the same figure)
+% 
+% The following input can be used instead of a fileName to create a
+% plot from some custom/pre-read GWAS summary file (ignored if fileName is 
+% provided):
+% pVals:        vector of p values (not -log10 values)
+%
+%% Check inputs
+% Check strTitle
+if ~exist('strTitle', 'var') || isempty(strTitle)
+    doTitle  = false;
+else
+    doTitle = true;
+end
+
+% Check h_ax
+if ~exist('h_ax', 'var') || isempty(h_ax)
+    createFig = true;
+else
+    if ishandle(h_ax)
+        createFig = false;
+    else
+        createFig = true;
+        warning('Invalid axis handle provided; ignoring h_ax as input');
+    end
+end
+
+%% Create figure
+if createFig
+    fig = figure('Units', 'centimeters', 'Position', [10 10 10 10]);
+    if doTitle
+        tight_subplot(1, 1, 0, [0.11 0.06], [0.1 0.02]);
+    else
+        tight_subplot(1, 1, 0, [0.11 0.01], [0.1 0.02]);
+    end
+    h_ax = gca;
+else
+    fig = h_ax.Parent;
+end
+
+hold(h_ax, 'on');
+
+% Define colour - 3-class Dark2 [colorbrewer2.org]
+if not(exist('colOver', 'var'))
+    colTrue  = [217, 95, 2]./255;
+else
+    colTrue = colOver;
+end
+
+% Create sorted data
+% Sort p values from lowest to highest
+p_observed = sort(pVals);
+
+% Largest -log10 p value
+maxVal = max(-log10(p_observed));
+
+% Create expected p values
+p_expected = (1:length(p_observed)) / length(p_observed);
+
+% Plot p values
+ll = plot(h_ax, -log10(p_expected), -log10(p_observed), '.', 'Color', colTrue);
+
+% Plot null
+plot(h_ax, [0, maxVal], [0, maxVal], 'Color', [0 0 0], 'LineStyle', '--', 'LineWidth', 1);
+
+% Round up maxVal
+maxVal = ceil(maxVal);
+
+% Customize axes limits
+xlim(h_ax, [0  maxVal+0.1]);
+ylim(h_ax, [0  maxVal+0.1]);
+
+% Place ticks and tick labels
+if maxVal > 6
+    xticks(h_ax,      0:2:maxVal);
+    yticks(h_ax,      0:2:maxVal);
+    xticklabels(h_ax, 0:2:maxVal);
+    yticklabels(h_ax, 0:2:maxVal);
+else
+    xticks(h_ax,      0:1:maxVal);
+    yticks(h_ax,      0:1:maxVal);
+    xticklabels(h_ax, 0:1:maxVal);
+    yticklabels(h_ax, 0:1:maxVal);
+end
+
+% Change font size for the tick labels
+h_ax.FontSize = 8;
+
+% Label the axes
+xlabel(h_ax, 'Expected -log_{10} \itp', 'Interpreter', 'tex', 'FontSize', 10, 'FontName', 'Courier');
+ylabel(h_ax, 'Observed -log_{10} \itp', 'Interpreter', 'tex', 'FontSize', 10, 'FontName', 'Courier');
+
+% Put a title, if user wants
+if doTitle
+    title(h_ax, strTitle, 'FontSize', 11, 'FontName', 'Courier');
+end